fix: move `SimpleApplication` to the cefsimple example per review
