+++ cef/Cargo.toml
@@ -44,7 +44,6 @@ windows = { workspace = true, optional = true }
 [target.'cfg(target_os = "macos")'.dependencies]
 libloading.workspace = true
 objc2.workspace = true
-objc2-app-kit = { workspace = true, features = ["NSApplication", "NSResponder"] }
 objc2-io-surface = { workspace = true, optional = true }
 objc = { workspace = true, optional = true }
 metal = { workspace = true, optional = true }
