+++ examples/cefsimple/Cargo.toml
@@ -25,3 +25,5 @@ cef-dll-sys.workspace = true
 [target.'cfg(target_os = "macos")'.dependencies]
 plist.workspace = true
 serde.workspace = true
+objc2.workspace = true
+objc2-app-kit = { workspace = true, features = ["NSApplication", "NSResponder"] }
